// src/utils/model.d.ts
/** @format */

declare namespace Model {
  // Generated by https://quicktype.io

  interface User {
    id: number
    nick_name: string
    avatar_url?: string
    country?: string
    province?: string
    city?: string
    gender?: number
    prefecture?: string
    post_count?: number
    followers_count: number
    following_count: number
    following: boolean
    summary: string
    wechat: string
    identity: string
    school: string
    labels: string
    company: string
    company_address: string
    email: string
    toutiao: string
    wechat_mp: string
    bilibili: string
    weibo: string
    cellphone: string
  }

  // Generated by https://quicktype.io

  interface Notification {
    id: number
    notify_title: string
    created_at: number
    notify_body: string
    notify_avatar_url: string
    post_id: string
    read: boolean
  }

  // Generated by https://quicktype.io

  interface UnreadCounts {
    likes_unread_count: number
    comments_unread_count: number
    system_unread_count: number
  }

  // Generated by https://quicktype.io

  interface Post {
    id: number
    body: string
    images_url: string[]
    grade: string
    created_at: number
    likes_count: number
    liked: boolean
    star: boolean
    user: User
    likes: User[]
    comments: any[]
    comments_count: number
    has_more_comments: boolean
    abilities: Actions
    thumbnails_url: string[]
  }

  // Generated by https://quicktype.io

  interface Comment {
    id: number
    body: string
    image_url?: string | null
    app_id: number
    created_at: number
    commentable_id: number
    reply_to?: Comment | null
    parent?: Comment | null
    children: Array<Comment>
    user: User
    abilities: Actions
  }


  // Generated by https://quicktype.io

  export interface Actions {
    destroy: boolean;
    top: boolean;
    untop: boolean;
    excellent: boolean;
    unexcellent: boolean;
  }

}
export default Model